requirements.txt
httpx>=0.27
python-dotenv>=1.0
pyyaml>=6.0
pytest>=8.0
pytest-asyncio>=0.23
respx>=0.21
ruff>=0.4
